APMech: The new (first) NetData API has been incorporated, and code has been adjusted accordingly. No functional changes (I hope).
Nabla-Wing: The ability to change nickname by a text command is
added. Some other minor adjustments.

NetData: The interface is polished a bit to unify the method
names, namespaces et cetera. Ogre detection is added too, so
ap::net::serializer can work with Ogre, but does not depend on it.
For the first time ever, both games share identical NetData code.

+++ src/server/Server.hpp
@@ -27,7 +27,7 @@ class Server
     uint32              mPort;
 
     void processEvents(ap::net::NetData *pNetData);
-    void updateObjects(uint64 dt, ap::net::NetData* pNetData) const;
+    void updateObjects(float dt, ap::net::NetData* pNetData) const;
     void fireWeapons(uint64 tstamp, ap::net::NetData *pNetData);
     void weaponFired(ap::net::NetData *pNetData, ap::MovingObject *source);
     void detectCollisions(ap::net::NetData *pNetData) const;
